u-boot: Add tie-test-builds and alternative key signing
Part of the testing for TI is to build alternative versions of uboot
that turn on/off certain features.  The decision was made to host these
config fragments in the ti-upstream repository so that they are  outside
of the tree, but centrally located for all version of u-boot to access.

This patch adds the logic to enable this feature by adding this to the
local.conf:

TI_EXTRAS += "tie-test-builds"

This makes use of the existing UBOOT_CONFIG flow for looping over uboot
and building it multiple times and naming the output files uniquely.

Additionally, we want to provide examples of how to sign the uboot files
with different keys so that customers can see how it is done.  We have
provided an ECDSA key and follow on patches will show to point to it.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>