u-boot-ti: Add k3/k3r5-falcon overrides
The two changes related to falcon for u-boot are building a tifalcon.bin
file in the a53 build, and applying a config fragment for the r5 builds.
To make sure we avoid releasing the same files as the non-falcon builds
we need to uniquely name all of the files produced by the r5 build, and
prune down the files for a53 builds.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>